Move updateHold to XCircService
